+++ test/datadog/gke_autopilot_allowlistedv2workload_test.go
@@ -64,7 +64,8 @@ func verifyDaemonsetAutopilotAllowlistedV2WorkloadMinimal(t *testing.T, manifest
 
 	assert.Equal(t, 1, len(ds.Spec.Template.Spec.Containers))
 	assert.Equal(t, ds.Spec.Template.Spec.Containers[0].Name, "agent")
-	requireEmptyDirVolume(t, ds, "dsdsocket")
+	requireNoVolume(t, ds, "dsdsocket")
+	requireNoEnvVar(t, ds, "DD_DOGSTATSD_SOCKET")
 
 	volumeNames := common.GetVolumeNames(ds)